# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Battery Research Workbench API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/experiments/{battery_id}/{experiment_id}/waveform-frames": {
      "get": {
        "operationId": "list_waveform_frames_api_v1_experiments__battery_id___experiment_id__waveform_frames_get",
        "parameters": [
          {
            "in": "path",
            "name": "battery_id",
            "required": true,
            "schema": {
              "title": "Battery Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Waveform Frames Api V1 Experiments  Battery Id   Experiment Id  Waveform Frames Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Waveform Frames",
        "tags": [
          "experiments",
          "runs"
        ]
      }
    },
    "/api/v1/experiments/{battery_id}/{experiment_id}/waveform-frames/{frame_index}": {
      "get": {
        "operationId": "get_waveform_frame_api_v1_experiments__battery_id___experiment_id__waveform_frames__frame_index__get",
        "parameters": [
          {
            "in": "path",
            "name": "battery_id",
            "required": true,
            "schema": {
              "title": "Battery Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "title": "Experiment Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "frame_index",
            "required": true,
            "schema": {
              "title": "Frame Index",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "max_points",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 1000,
              "minimum": 1,
              "title": "Max Points",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Waveform Frame Api V1 Experiments  Battery Id   Experiment Id  Waveform Frames  Frame Index  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Waveform Frame",
        "tags": [
          "experiments",
          "runs"
        ]
      }
    },
...
    "/api/v1/runs": {
      "get": {
        "operationId": "list_runs_api_v1_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Runs Api V1 Runs Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "experiments",
          "runs"
        ]
      },
      "post": {
        "operationId": "start_run_api_v1_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Start Run Api V1 Runs Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Run",
        "tags": [
          "runs"
        ]
      }
    },
...
  },
...
}